Move pkg/config/cmd to pkg/cmd/util
