+++ pkg/cmd/cli/cmd/newbuild.go
@@ -17,7 +17,6 @@ import (
 	buildapi "github.com/openshift/origin/pkg/build/api"
 	cmdutil "github.com/openshift/origin/pkg/cmd/util"
 	"github.com/openshift/origin/pkg/cmd/util/clientcmd"
-	configcmd "github.com/openshift/origin/pkg/config/cmd"
 	newapp "github.com/openshift/origin/pkg/generate/app"
 	newcmd "github.com/openshift/origin/pkg/generate/app/cmd"
 )
@@ -82,7 +81,7 @@ on the Docker Hub.
 )
 
 type NewBuildOptions struct {
-	Action configcmd.BulkAction
+	Action cmdutil.BulkAction
 	Config *newcmd.AppConfig
 
 	BaseName    string
@@ -162,7 +161,7 @@ func (o *NewBuildOptions) Complete(baseName, commandName string, f *clientcmd.Fa
 
 	o.Action.Out, o.Action.ErrOut = o.Out, o.ErrOut
 	o.Action.Bulk.Mapper = clientcmd.ResourceMapper(f)
-	o.Action.Bulk.Op = configcmd.Create
+	o.Action.Bulk.Op = cmdutil.Create
 	// Retry is used to support previous versions of the API server that will
 	// consider the presence of an unknown trigger type to be an error.
 	o.Action.Bulk.Retry = retryBuildConfig
@@ -219,7 +218,7 @@ func (o *NewBuildOptions) RunNewBuild() error {
 		return o.PrintObject(result.List)
 	}
 
-	if errs := o.Action.WithMessage(configcmd.CreateMessage(config.Labels), "created").Run(result.List, result.Namespace); len(errs) > 0 {
+	if errs := o.Action.WithMessage(cmdutil.CreateMessage(config.Labels), "created").Run(result.List, result.Namespace); len(errs) > 0 {
 		return cmdutil.ErrExit
 	}
