+++ pkg/cmd/util/config_test.go
@@ -1,4 +1,4 @@
-package cmd
+package util
 
 import (
 	"bytes"
